truck: rename file and update imports
