+++ src/modules/scheduler/interfaces/scheduler.interface.ts
